Query user for DB connection details.
Right now this leaks EventEmitters, but I think this might be a bug in Prompt, Read, or perhaps readline.
It looks like my usage isn't causing readline.close() to be called which is leaving listeners open.
Read's done() should call this, but I'm not sure what should be calling done(), so I need to trace this.